+++ apisix/ssl/router/radixtree_sni.lua
@@ -25,6 +25,7 @@ local type             = type
 local error            = error
 local str_find         = core.string.find
 local str_gsub         = string.gsub
+local str_lower        = string.lower
 local ssl_certificates
 local radixtree_router
 local radixtree_router_ver
@@ -226,6 +227,21 @@ function _M.ssls()
 end
 
 
+local function ssl_filter(ssl)
+    if not ssl.value then
+        return
+    end
+
+    if ssl.value.sni then
+        ssl.value.sni = str_lower(ssl.value.sni)
+    elseif ssl.value.snis then
+        for i, v in ipairs(ssl.value.snis) do
+            ssl.value.snis[i] = str_lower(v)
+        end
+    end
+end
+
+
 function _M.init_worker()
     local err
     ssl_certificates, err = core.config.new("/ssl", {
@@ -234,6 +250,7 @@ function _M.init_worker()
         checker = function (item, schema_type)
             return apisix_ssl.check_ssl_conf(true, item)
         end,
+        filter = ssl_filter,
     })
     if not ssl_certificates then
         error("failed to create etcd instance for fetching ssl certificates: "