+++ ci/linux_apisix_master_luarocks_runner.sh
@@ -36,6 +36,8 @@ script() {
     mkdir tmp && cd tmp
     cp -r ../utils ./
 
+    git config --global url.https://github.com/.insteadOf git://github.com/
+
     # install APISIX by luarocks
     sudo luarocks install $APISIX_MAIN > build.log 2>&1 || (cat build.log && exit 1)
     cp ../bin/apisix /usr/local/bin/apisix